% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishstatr.R
\docType{package}
\name{fishstatr}
\alias{fishstatr}
\alias{fishstatr-package}
\title{fishstatr: an R package to access fishstat reference data.}
\description{
The fishstatr package provides fishstat reference
data from EBX5.
This package depends on the EBX5 folder called 'Metadata', containing
the 'FistStat' data structure in place, and populated with data.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}

Luís G. Silva e Silva, \email{luis.silvaesilva@fao.org}
}
