% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebx5_GetCodelistByID.R
\name{GetCodelistByID}
\alias{GetCodelistByID}
\title{Get en EBX5 codelist using the ID}
\usage{
GetCodelistByID(ebxCodeListID)
}
\arguments{
\item{ebxCodeListID}{EBX5 codelist ID}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the EBX5 codelist for a given EBX5 codelist ID;
codelists are normally referred by Acronymn, but fishstat uses the ID
This function requires that EBX5 metadata was loded already
}
\examples{

\dontrun{
GetCodelistByID(200)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
