% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetConcepts.R
\name{GetConcepts}
\alias{GetConcepts}
\title{Get FishStat Concepts}
\usage{
GetConcepts(metadata, datasetID = 1)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using ReadMetadata()}

\item{datasetID}{dataset Identifier.}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the concept IDs for a given dataset
}
\examples{

\dontrun{
GetConcepts()
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
