% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_GetDatasetConcepts.R
\name{GetDatasetConcepts}
\alias{GetDatasetConcepts}
\title{Get FishStat Dataset Concepts}
\usage{
GetDatasetConcepts(metadata, datasetID)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}

\item{datasetID}{dataset Identifier obtained using \code{\link{GetDatasets}}}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
of a single column listing the FishStat ConceptIDs.
}
\description{
This function returns the concept IDs for a given dataset
}
\examples{

\dontrun{
metadata <- ReadMetadata()
GetDatasetConcepts(metadata, datasetID = 1)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
