% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetDatasetDimensions.R
\name{GetDatasetDimensions}
\alias{GetDatasetDimensions}
\title{Get FishStat Dataset Dimensions}
\usage{
GetDatasetDimensions(metadata, datasetID)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}

\item{datasetID}{dataset Identifier obtained using \code{\link{GetDatasets}}}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the Dimensions for a given dataset.
A Dimension is based on a Concept(Country). The field of this concept used for the
dimension is called Attribute(UN_Code).
The EBXCodelist-ID is what hierarchies for this dimension are based on.
}
\examples{

\dontrun{
metadata <- ReadMetadata()
GetDatasetDimensions(metadata, datasetID = 1)
   AttributeID ConceptID DimensionID                Name_En EBXCodelist    EBXName
1:           5         1          11                Country         200    UN_Code
2:         101         2          12          ASFIS species         301 Alpha_Code
3:          41         8          13 FAO major fishing area         403       Code
4:         151        30          14            Environment         502       Code
}

}
\seealso{
\code{\link{GetEBXHierarchy}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
