% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetDatasets.R
\name{GetDatasets}
\alias{GetDatasets}
\title{Get FishStat Datasets}
\usage{
GetDatasets(metadata)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This is a convenience function to make it easier to get started usign data
from EBX5. It provides a dataset name, and the Dataset-Identifier required by other fucntions.
}
\examples{

\dontrun{
library(faoebx5)
library(fishstatr)
metadata <- ReadMetadata()
GetDatasets(metadata)
   Identifier           Acronym                                               Name_En
1            1       AQUACULTURE                         Global aquaculture production
2            2           CAPTURE                             Global capture production
3            3 GLOBAL_PRODUCTION                Global production by production source
4            4             TRADE            Fisheries commodities production and trade
5            5     CECAF_CAPTURE   CECAF (Eastern Central Atlantic) capture production
6            6      GFCM_CAPTURE GFCM (Mediterranean and Black Sea) capture production
7            7    RECOFI_CAPTURE                             RECOFI capture production
8            8     SEATL_CAPTURE                 Southeast Atlantic capture production
9            9      FOOD_BALANCE      Food balance sheets of fish and fishery products
10          10        POPULATION                                      World population
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
