% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_GetDimensionGroupData.R
\name{GetDimensionGroupData}
\alias{GetDimensionGroupData}
\title{Get FishStat Dimension Data}
\usage{
GetDimensionGroupData(metadata, dimensionConceptID, ebxConceptID)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}

\item{dimensionConceptID}{dimension Identifier (a concept)}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the groups for a given dimension (Concept) inclduing the child codes
}
\examples{

\dontrun{
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
