% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetDimensionGroups.R
\name{GetDimensionGroups}
\alias{GetDimensionGroups}
\title{Get FishStat Dimension}
\usage{
GetDimensionGroups(metadata, dimensionConceptID = 1)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using ReadMetadata()}

\item{dimensionConceptID}{dimension Identifier (a concept)}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the groups for a given dimension (Concept)
}
\examples{

\dontrun{
GetConcepts()
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
