% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetDimensions.R
\name{GetDimensions}
\alias{GetDimensions}
\title{Get FishStat Dimensions}
\usage{
GetDimensions(metadata, datasetID = 1)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using ReadMetadata()}

\item{datasetID}{dataset Identifier.}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the AttributeKeys for a given dataset.
AttributeKey is a EBX5 composite key, in the form of paste0(ConceptID,'|',AttributeID)
}
\examples{

\dontrun{
GetDimension(datasetID=2)
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
