% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_GetGroupConnections.R
\name{GetGroupConnections}
\alias{GetGroupConnections}
\title{Get the connections between two hierarchies}
\usage{
GetGroupConnections(ebx_parent_group, ebx_child_group)
}
\arguments{
\item{ebx_parent_group}{parent, using the EBX5 codelist Identifier}

\item{ebx_child_group}{child, using the EBX5 codelist Identifier}

\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}
}
\value{
grouping in the format int=parent list=(child1,child2, child3...)
}
\description{
Using the EBX5 metadata, calculates the grouping from parent to child group,
based on codelist Identifier; as defined in 'EBXGroup.to' and 'EBXGroup.from'.
This is a basic function of this package, it is exposed primarily for troubleshooting.
Users will normally use \code{\link{GetEBXHierarchy}}.
}
\details{
The solution may invaolve several steps (a list of group ID's), and may involve
multiple step groupings, which have to be merged for the final result:
   MAJOR(306)->ODER(307)->FAMILY(302)->SPECIES(301)
   MAJOR(306)->ODER(307)->SPECIES(301)
In the simplest case, if a matching group exists,
list(ebx_parent_group, ebx_child_group) is returned
}
\examples{
\dontrun{
GetGroupConnections(306,301)
[[1]]
[1] "306" "307" "302" "301"
[[2]]
[1] "306" "307" "301"
}

}
\seealso{
\code{\link{[faoebx5]ReadEBXGroup}} \code{\link{GetDimensionGroups}}
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
