% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_GroupAsList.R
\name{GroupAsList}
\alias{GroupAsList}
\title{Convert an grouping from EBX5 to a list}
\usage{
GroupAsList(grouping)
}
\arguments{
\item{grouping}{a grouping in the format (int=parent, int=child), as returned by EBX5}
}
\value{
grouping in the format int=parent list=(child1,child2, child3...)
}
\description{
This function will merge the child ID's belonging to a single parent ID
}
\examples{
\dontrun{
df <- data.frame(group=c(100,100, 200, 200), children=c(10,11,21,22))
str(GroupAsList(df))
 $ grouping[, 1]: num  100 200
 $ children     :List of 2
   ..$ : num  10 11
   ..$ : num  21 22

library(faoebx5)
df <- ReadEBXGroup('HCL_FI_COUNTRY_CONTINENT_GEOREGION')
GroupAsList(df)
}


}
\seealso{
\code{\link{ReadEBXGroup}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
