% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_ReadMetadata.R
\name{ReadMetadata}
\alias{ReadMetadata}
\title{Read FishStat Metadata}
\usage{
ReadMetadata()
}
\value{
The returned object is used by other methods
}
\description{
This function returns the FishStat metadata which is used for subsequent operations
}
\examples{

\dontrun{
GetConcepts()
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
