% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_fish.R
\name{rebuild_fish}
\alias{rebuild_fish}
\title{Assembles FAO's Global Fishery and Aquaculture Production Data from scratch}
\usage{
rebuild_fish(path_to_zipfile)
}
\arguments{
\item{path_to_zipfile}{Name of zipped file of FAO data (if in current directory); otherwise, specify path to file}
}
\value{
A merged, tidy dataset.
}
\description{
\code{rebuild_fish} Extracts data from FAO zipped file and merges them into "tidy" (i.e., long) format
}
\examples{
rebuild_fish("~/OneDrive - american.edu/GlobalProduction_2019.1.0.zip")
rebuild_fish("C:\\\\Users\\\\kdgor\\\\OneDrive - american.edu\\\\GlobalProduction_2019.1.0.zip")
}
