% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fish.R
\name{clean_fish}
\alias{clean_fish}
\title{Takes downloaded FishStatJ Global Fishery and Aquaculture Production data, cleans the time series data and makes it "tidy" (i.e., long format).}
\usage{
clean_fish(path_to_file)
}
\arguments{
\item{path_to_file}{Path to FAO data}
}
\value{
A cleaned, tidy dataset
}
\description{
Takes downloaded FishStatJ Global Fishery and Aquaculture Production data, cleans the time series data and makes it "tidy" (i.e., long format).
}
