% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_GetDimensionGroups.R
\name{GetDimensionGroups}
\alias{GetDimensionGroups}
\title{Get FishStat Dimension Groups}
\usage{
GetDimensionGroups(metadata, dimensionConceptID)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}

\item{dimensionConceptID}{dimension Identifier (a concept) obtained using \code{\link{GetDimensions}}}
}
\value{
Returns an object of the class \code{\link[data.table]{data.table}}
}
\description{
This function returns the groups for a given dimension (Concept).
On a dataset with dimension country, the dimension groups (or hierachies) are used for
aggregation (Continent, Region, ...).
}
\examples{

\dontrun{
metadata <- ReadMetadata()
GetDimensionGroups(metadata, dimensionConceptID = 1)
   Identifier             Acronym Sort EBXCodelist                                Name_En
3           3           CONTINENT  102         201                              Continent
4           4          GEO_REGION  103         208                    Geographical region
5           5          ECON_CLASS  105         202                         Economic class
6           6          ECON_GROUP  106         203                         Economic group
}

}
\seealso{
\code{\link{GetEBXHierarchy}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
