% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ReadDatasetCodelists.R
\name{ReadDatasetCodelists}
\alias{ReadDatasetCodelists}
\title{Read a complete group hierarchy, for use with applications}
\usage{
ReadDatasetCodelists(metadata, datasetID)
}
\arguments{
\item{metadata}{FishStat metadata; obtained using \code{\link{ReadMetadata}}}

\item{datasetID}{dataset Identifier obtained using \code{\link{GetDatasets}}}
}
\description{
This function will read all codelists and hierarchies
}
\details{
It first calls \code{\link{GetDatasetDimensions}} to obtain the dimensions for
the dataset. For each of the dimensions, calls \code{\link{GetDimensionGroups}} to
obtain the hierarchies for this dimension. Finally calls \code{\link{ReadEBXHierarchy}}
to obtain the grouping from the bottom hierarchy (as specified in GetDatasetDimensions) to the
top (as specified in GetDimensionGroups).
It saves all objects into an RData file with the name of the dataset.
}
\examples{
\dontrun{
 library(faoebx5)
 library(fishstatr)
 #use row1 for this example
 datasetID <- GetDatasets(metadata)[1,'Identifier']
 datasetName <- GetDatasets(metadata)[1,'Acronym']

 ReadDatasetCodelists(metadata, datasetID)
 [1] "=== saved 29 to AQUACULTURE.RData, size=349984"

 load(file = paste0(datasetName,'.RData'))
 > ls()
[1] "COUNTRY.Codelist"                   "COUNTRY.Groups"                     "COUNTRY.Groups.CONTINENT"
[4] "COUNTRY.Groups.GEO_REGION"          "COUNTRY.Groups.ECON_CLASS"          "COUNTRY.Groups.ECON_GROUP"
[7] "COUNTRY.Groups.COMMISSION"          "COUNTRY.Groups.ECO_REGION"          "COUNTRY.Groups.OTHER_COUNTRY_GROUP"
[10] "SPECIES.Codelist"                   "SPECIES.Groups"                     "SPECIES.Groups.YEARBOOK_GROUP"
[13] "SPECIES.Groups.ISSCAAP_DIVISION"    "SPECIES.Groups.ISSCAAP_GROUP"       "SPECIES.Groups.MAIN_GROUP"
[16] "SPECIES.Groups.ORDER"               "SPECIES.Groups.FAMILY"              "SPECIES.Groups.SPECIES_GROUP"
[19] "SPECIES.Groups.CPC_DIVISION"        "SPECIES.Groups.CPC_GROUP"           "SPECIES.Groups.CPC"
[22] "AREA.Codelist"                      "AREA.Groups"                        "AREA.Groups.INLAND_MARINE"
[25] "AREA.Groups.OCEAN"                  "AREA.Groups.SUB_OCEAN"              "AREA.Groups.FA_REGION"
[28] "ENVIRONMENT.Codelist"               "ENVIRONMENT.Groups"
}


}
\seealso{
\code{\link{GetDimensionGroups}} and \code{\link{ReadEBXHierarchy}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
