% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ReadEBXHierarchy.R
\name{ReadEBXHierarchy}
\alias{ReadEBXHierarchy}
\title{Read a comple group hierarchy}
\usage{
ReadEBXHierarchy(sdmx_level1_ID, sdmx_level2_ID)
}
\arguments{
\item{sdmx_level1}{ID upper level of a hierarchy, as returned by GetDimensionGroups$EBXCodelist}

\item{sdmx_level2}{lower level of a hierarchy, as returned by GetDimensions$EBXCodelist}
}
\value{
grouping in the format int=parent list=(child1,child2,child3...)
}
\description{
This function will resolve all possible groupings, accross multiple levels,
 combining these levels into the final result. In case there are multiple group paths
 between the levels requested, all paths are resolved and combined.
}
\details{
The group names are using the EBX5-ID, returned by
}
\examples{
\dontrun{
ReadEBXHierarchy(306,301)
}


}
\seealso{
\code{\link{GetDimensionGroups}} and \code{\link{GetDatasetDimensions}}.
}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
