% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_ReadMetadata.R
\name{ReadMetadata}
\alias{ReadMetadata}
\title{Read FishStat Metadata}
\usage{
ReadMetadata()
}
\value{
The returned object is used by other methods
}
\description{
This function returns the FishStat metadata which is used for subsequent operations
}
\examples{

\dontrun{
install.packages("C:/Programs/GithubTom/faoebx5_1.0.0.tar.gz", repos = NULL, type="source")
devtools::install_github('bergertom/faoebx5')
library(faoebx5)
SetupEBXConnection('Fishery','Fishery','http://server.name/ebx-dataservices/connector','user', 'pass')
metadata <- ReadMetadata()
}

}
\author{
Thomas Berger, \email{thomas.berger@fao.org}
}
