% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_fish.R
\name{rebuild_fish}
\alias{rebuild_fish}
\title{Assembles FAO's Global Fishery and Aquaculture Production Data from scratch}
\usage{
rebuild_fish(path_to_folder)
}
\arguments{
\item{path_to_folder}{Path to unzipped (extracted) folder of FAO data}
}
\value{
A merged, tidy dataset.
}
\description{
\code{rebuild_fish} takes all CSV files from an unzipped folder of FAO data and merges them into "tidy" (i.e., long) format
}
\examples{
clean_fish("~/OneDrive - american.edu/GlobalProduction_2019.1.0")
clean_fish("C:\\\\Users\\\\kdgor\\\\OneDrive - american.edu\\\\GlobalProduction_2019.1.0")
}
